Derive Macro git_function_history_proc_macro::enumstuff

#[derive(enumstuff)]
{
    // Attributes available to this derive:
    #[enumstuff]
}

Allows the type that derives this macro, to have a method from_str that takes a list of strings and returns the type.

use [enumstuff(skip)] attribute on a variant or field to make it not able to be parsed by from_str.